[fix] 홈 폴더 생성 viewModel 바꾸기
