--- feature/home/src/main/java/com/mashup/dorabangs/feature/createfolder/HomeCreateFolderSideEffect.kt
@@ -1,5 +1,0 @@
-package com.mashup.dorabangs.feature.createfolder
-
-sealed class HomeCreateFolderSideEffect {
-    object NavigateToHome : HomeCreateFolderSideEffect()
-}
