--- feature/home/src/main/java/com/mashup/dorabangs/feature/createfolder/HomeCreateFolderState.kt
@@ -1,7 +1,0 @@
-package com.mashup.dorabangs.feature.createfolder
-
-data class HomeCreateFolderState(
-    val folderName: String = "",
-    val helperEnable: Boolean = false,
-    val helperMessage: String = "",
-)
