+++ feature/home/src/main/java/com/mashup/dorabangs/feature/home/HomeSideEffect.kt
@@ -4,4 +4,5 @@ sealed class HomeSideEffect {
     data class ShowSnackBar(val copiedText: String) : HomeSideEffect()
     object HideSnackBar : HomeSideEffect()
     object NavigateToCreateFolder : HomeSideEffect()
+    object NavigateToHome : HomeSideEffect()
 }
